chore(hogql): move clickhouse branches out of BasePrinter
Drains every `self.dialect == "clickhouse"` check from the shared
BasePrinter and into ClickHousePrinter. After this change, BasePrinter
contains zero CH-specific branches; ClickHousePrinter owns all CH
rendering logic.

- visit_call: refactored into a ~30-line dispatch that delegates to
  `_render_aggregation_call`, `_render_function_call`, and
  `_render_posthog_function_call` hooks. The ~200-line CH function
  rendering (FIRST_ARG_DATETIME_FUNCTIONS, concat, suffix_args,
  overloads, tz_aware, toStartOfWeek mode, trim*, etc.) moved
  verbatim into ClickHousePrinter overrides. Same for the posthog
  function renderings (embedText, lookupDomainType, convertCurrency, …).
- LIMIT percent / WITH TIES: now a `_render_select_query_limit_clause`
  hook (plus `_render_set_query_limit_percent` for set queries); CH
  and PG override with their respective behaviors. Base raises for
  unsupported dialects.
- Aggregation rendering: `_render_aggregation_name` hook keeps the
  ClickHouse name by default; HogQLPrinter overrides to preserve
  `node.name` for HogQL output. `_validate_within_group_for_aggregation`
  hook lets CH reject WITHIN GROUP.
- Property groups materialization: `_yield_property_group_columns`
  hook; default yields nothing, CH yields `PrintableMaterializedPropertyGroupItem`s.
- `_min_function_name` hook (default `"min2"`, PG overrides with
  `"least"`) replaces the inline PG branch at base.py:313.
- Removed the dead `self.dialect == "hogql"` check at clickhouse.py:875.
- Moved CH-only imports (re, django_settings, ADD_OR_NULL_DATETIME_FUNCTIONS,
  FIRST_ARG_DATETIME_FUNCTIONS, resolve_embed_text, EXCHANGE_RATE_DICTIONARY_NAME,
  Constant, StringType, property_groups, PrintableMaterializedPropertyGroupItem)
  out of base.py.

No behavior change — all 521 printer tests pass across HogQL, ClickHouse,
and Postgres dialects.
